requirements.txt
Flask
flask-cors
flask-sqlalchemy
pydantic
sentence-transformers
faiss-cpu
sqlalchemy
aiosqlite